test: consolidate suite to 13 files and raise coverage to 97%
Consolidate 25 test files into 13 domain-focused files: dissolve the
print-smoke and line-targeted coverage files (test_all_objects.py,
test_coverage_extra.py), delete duplicated suites (test_object_group.py,
test_group_behavior.py), and merge small files into their domains
(sensors, gui, plot, behaviors, world_map, objects, env). Ported tests
are rewritten with real assertions; loop-behavior YAML fixtures are
inlined via tmp_path.

Measure coverage of irsim only: pytest/coverage config moves into
pyproject.toml (--cov=irsim by default) and CI drops --cov . so codecov
no longer blends tests/ and usage/ into the number. Add targeted tests
for previously uncovered paths (lazy irsim.lib exports, 3D geometry
handler, behavior goal-None branches, collision-mode edge cases,
deprecated robot shims, World3D offsets, backend fallback), and
re-enable the F5 debug test that was skipped due to the global state
pollution fixed in the config proxy commit.

Coverage of irsim/ rises from 96.7% to 97.3% with ~2400 fewer test
lines and a ~30% faster suite.
