--- tests/test.md
@@ -1,1 +1,0 @@
-pytest --cov=ir-sim --cov-report html
