--- tests/test_loop_multiple.yaml
@@ -1,11 +1,0 @@
-world:
-  width: 10
-  height: 10
-
-robot:
-  kinematics: {name: diff}
-  state: [1, 1, 0]
-  goal: [[3, 3, 0], [5, 5, 0], [7, 7, 0]]
-  behavior:
-    name: dash
-    loop: true
