--- tests/test_loop_omni.yaml
@@ -1,11 +1,0 @@
-world:
-  width: 10
-  height: 10
-
-robot:
-  kinematics: {name: omni}
-  state: [1, 1, 0]
-  goal: [[3, 3], [5, 5]]
-  behavior:
-    name: dash
-    loop: true
