--- tests/test_loop_single.yaml
@@ -1,11 +1,0 @@
-world:
-  width: 10
-  height: 10
-
-robot:
-  kinematics: {name: diff}
-  state: [1, 1, 0]
-  goal: [5, 5]
-  behavior:
-    name: rvo
-    loop: true
